refactor: Cleanup component mappings and utils (python side) (#523)
This is the first of 2 PRs to simplify our DH UI component mappings
since we no longer have the Spectrum dependency. This one addresses the
python modules + minimal re-mapping in the JS to make it still work. The
next one will cleanup the .js side more holistically.

* Removed components/spectrum subfolder and moved component modules to
parent
* Moved types into a components/types folder
* Minimal changes to SpectrumUtils to make mapping still work
* Added a broadstrokes e2e test to prove all components render

resolves #425
