+++ plugins/ui/src/deephaven/ui/components/__init__.py
@@ -1,26 +1,52 @@
+from .action_button import action_button
 from .action_group import action_group
 from .action_menu import action_menu
-from .icon import icon
-from .make_component import make_component as component
-from .fragment import fragment
-from .panel import panel
-from .spectrum import *
-from .table import table
-from .dashboard import dashboard
-from .row import row
+from .basic import (
+    component_element,
+    grid,
+    heading,
+    icon_wrapper,
+    illustrated_message,
+    form,
+    switch,
+    tabs,
+    tab_list,
+    tab_panels,
+    text,
+    view,
+)
+from .button import button
+from .button_group import button_group
+from .checkbox import checkbox
 from .column import column
-from .stack import stack
-from .picker import picker
-from .action_group import action_group
-from .section import section
-from .action_menu import action_menu
+from .combo_box import combo_box
+from .content import content
+from .contextual_help import contextual_help
+from .dashboard import dashboard
+from .date_picker import date_picker
+from .flex import flex
+from .fragment import fragment
+from .icon import icon
 from .item import item
-from .list_view import list_view
+from .item_table_source import item_table_source
 from .list_action_group import list_action_group
 from .list_action_menu import list_action_menu
-from .item_table_source import item_table_source
+from .list_view import list_view
+from .make_component import make_component as component
+from .number_field import number_field
+from .panel import panel
+from .picker import picker
 from .radio import radio
 from .radio_group import radio_group
+from .range_slider import range_slider
+from .row import row
+from .section import section
+from .slider import slider
+from .stack import stack
+from .table import table
+from .text_field import text_field
+from .toggle_button import toggle_button
+from .types import *
 
 from . import html
 
@@ -29,14 +55,17 @@ __all__ = [
     "action_button",
     "action_group",
     "action_menu",
+    "component_element",
     "button",
     "button_group",
     "checkbox",
     "column",
+    "combo_box",
     "component",
     "content",
     "contextual_help",
     "dashboard",
+    "date_picker",
     "flex",
     "form",
     "fragment",
@@ -60,7 +89,6 @@ __all__ = [
     "row",
     "section",
     "slider",
-    "spectrum_element",
     "stack",
     "switch",
     "table",
