+++ plugins/ui/src/deephaven/ui/components/action_group.py
@@ -3,14 +3,12 @@ from numbers import Number
 from typing import Any, Callable, Iterable
 
 
-from ..components.spectrum.events import (
+from .types import (
+    # Events
     ButtonLabelBehavior,
     Orientation,
     StaticColor,
-)
-from ..elements import Element, BaseElement
-from ..types import ActionGroupDensity, SelectedKeys, SelectionMode, Key, Selection
-from .spectrum.layout import (
+    # Layout
     AlignSelf,
     CSSProperties,
     DimensionValue,
@@ -19,6 +17,9 @@ from .spectrum.layout import (
     OverflowMode,
     Position,
 )
+from .basic import component_element
+from ..elements import Element
+from ..types import ActionGroupDensity, SelectedKeys, SelectionMode, Key, Selection
 
 
 def action_group(
@@ -153,8 +154,8 @@ def action_group(
         UNSAFE_class_name: Set the CSS className for the element. Only use as a last resort. Use style props instead.
         UNSAFE_style: Set the inline style for the element. Only use as a last resort. Use style props instead.
     """
-    return BaseElement(
-        "deephaven.ui.components.ActionGroup",
+    return component_element(
+        "ActionGroup",
         *children,
         is_emphasized=is_emphasized,
         density=density,
