+++ plugins/ui/src/deephaven/ui/components/action_menu.py
@@ -5,11 +5,20 @@ from typing import Callable, Iterable
 from .item import Item
 from .section import SectionElement
 
-from .spectrum.events import TriggerType
 from ..types import Key, ActionKey, ActionMenuDirection
-from ..elements import BaseElement, Element
+from .basic import component_element
+from ..elements import Element
 
-from .spectrum import (
+from ..types import (
+    Key,
+    ActionKey,
+    ActionMenuDirection,
+)
+
+from .types import (
+    # Events
+    TriggerType,
+    # Layout
     Alignment,
     AlignSelf,
     CSSProperties,
@@ -141,8 +150,8 @@ def action_menu(
         UNSAFE_class_name: Set the CSS className for the element. Only use as a last resort. Use style props instead.
         UNSAFE_style: Set the inline style for the element. Only use as a last resort. Use style props instead.
     """
-    return BaseElement(
-        f"deephaven.ui.components.ActionMenu",
+    return component_element(
+        f"ActionMenu",
         *children,
         is_disabled=is_disabled,
         is_quiet=is_quiet,
