+++ plugins/ui/src/deephaven/ui/components/content.py
@@ -1,15 +1,7 @@
 from __future__ import annotations
 from typing import Any, Callable
-from .accessibility import AriaExpanded, AriaHasPopup, AriaPressed
-from .events import (
-    ButtonType,
-    FocusEventCallable,
-    KeyboardEventCallable,
-    PressEventCallable,
-    StaticColor,
-    Orientation,
-)
-from .layout import (
+from .types import (
+    # Layout
     AlignSelf,
     CSSProperties,
     DimensionValue,
@@ -17,33 +9,13 @@ from .layout import (
     LayoutFlex,
     Number,
     Position,
-    LabelPosition,
 )
-from .basic import spectrum_element
-from ...elements import Element
+from .basic import component_element
+from ..elements import Element
 
 
-def checkbox(
+def content(
     *children: Any,
-    is_emphasized: bool | None = None,
-    is_indeterminate: bool | None = None,
-    default_selected: bool | None = None,
-    is_selected: bool | None = None,
-    value: str | None = None,
-    is_disabled: bool | None = None,
-    is_read_only: bool | None = None,
-    is_required: bool | None = None,
-    is_invalid: bool | None = None,
-    # validation_behaviour, # omitted because validate is not implemented
-    # validate, # omitted because it needs to return a ValidationError synchronously
-    auto_focus: bool | None = None,
-    name: str | None = None,
-    on_change: Callable[[bool], None] | None = None,
-    on_focus: FocusEventCallable | None = None,
-    on_blur: FocusEventCallable | None = None,
-    on_focus_change: Callable[[bool], None] | None = None,
-    on_key_down: KeyboardEventCallable | None = None,
-    on_key_up: KeyboardEventCallable | None = None,
     flex: LayoutFlex | None = None,
     flex_grow: Number | None = None,
     flex_shrink: Number | None = None,
@@ -81,38 +53,14 @@ def checkbox(
     z_index: Number | None = None,
     is_hidden: bool | None = None,
     id: str | None = None,
-    exclude_from_tab_order: bool | None = None,
-    aria_controls: str | None = None,
-    aria_label: str | None = None,
-    aria_labelledby: str | None = None,
-    aria_describedby: str | None = None,
-    aria_details: str | None = None,
-    aria_errormessage: str | None = None,
     UNSAFE_class_name: str | None = None,
     UNSAFE_style: CSSProperties | None = None,
 ) -> Element:
     """
-    Checkboxes allow users to select multiple items from a list of individual items, or to mark one individual item as selected.
+    Content represents the primary content within a Spectrum container.
 
     Args:
-        children: The checkbox label.
-        is_emphasized: This prop sets the emphasized style which provides visual prominence.
-        is_indeterminate: Indeterminism is presentational only. The indeterminate visual representation remains regardless of user interaction.
-        default_selected: Whether the element should be selected (uncontrolled).
-        is_selected: Whether the element should be selected (controlled).
-        value: The value of the input element, used when submitting a form.
-        is_disabled: Whether the element is disabled.
-        is_read_only: Whether the element is read-only.
-        is_required: Whether the element is required before form submission.
-        is_invalid: Whether the element is invalid.
-        auto_focus: Whether the element should automatically get focus on render.
-        name: The name of the input element, used when submitting a form.
-        on_change: Handler that is called when the element is selected or deselected.
-        on_focus: Handler that is called when the element receives focus.
-        on_blur: Handler that is called when the element loses focus.
-        on_focus_change: Handler that is called when the element receives or loses focus.
-        on_key_down: Handler that is called when a key is pressed down.
-        on_key_up: Handler that is called when a key is released.
+        children: The content to render within the container.
         flex: When used in a flex layout, specifies how the element will grow or shrink to fit the space available.
         flex_grow: When used in a flex layout, specifies how the element will grow to fit the space available.
         flex_shrink: When used in a flex layout, specifies how the element will shrink to fit the space available.
@@ -150,39 +98,12 @@ def checkbox(
         z_index: The stack order of the element.
         is_hidden: Whether the element is hidden.
         id: The unique identifier of the element.
-        exclude_from_tab_order: Whether the element should be excluded from the tab order. If true, the element will not be focusable via the keyboard by tabbing.
-        aria_controls: The id of the element that the current element controls.
-        aria_label: The label for the element.
-        aria_labelledby: The id of the element that labels the current element.
-        aria_describedby: The id of the element that describes the current element.
-        aria_details: The id of the element that provides additional information about the current element.
-        aria_errormessage: The id of the element that provides error information for the current element.
         UNSAFE_class_name: A CSS class to apply to the element.
         UNSAFE_style: A CSS style to apply to the element.
     """
-
-    return spectrum_element(
-        "Checkbox",
+    return component_element(
+        "Content",
         children=children,
-        is_emphasized=is_emphasized,
-        is_indeterminate=is_indeterminate,
-        default_selected=default_selected,
-        is_selected=is_selected,
-        value=value,
-        is_disabled=is_disabled,
-        is_read_only=is_read_only,
-        is_required=is_required,
-        is_invalid=is_invalid,
-        # validation_behaviour = validation_behaviour,
-        # validate = validate,
-        auto_focus=auto_focus,
-        name=name,
-        on_change=on_change,
-        on_focus=on_focus,
-        on_blur=on_blur,
-        on_focus_change=on_focus_change,
-        on_key_down=on_key_down,
-        on_key_up=on_key_up,
         flex=flex,
         flex_grow=flex_grow,
         flex_shrink=flex_shrink,
@@ -220,13 +141,6 @@ def checkbox(
         z_index=z_index,
         is_hidden=is_hidden,
         id=id,
-        exclude_from_tab_order=exclude_from_tab_order,
-        aria_controls=aria_controls,
-        aria_label=aria_label,
-        aria_labelledby=aria_labelledby,
-        aria_describedby=aria_describedby,
-        aria_details=aria_details,
-        aria_errormessage=aria_errormessage,
         UNSAFE_class_name=UNSAFE_class_name,
         UNSAFE_style=UNSAFE_style,
     )
