+++ plugins/ui/src/deephaven/ui/components/contextual_help.py
@@ -1,30 +1,33 @@
 from __future__ import annotations
+from numbers import Number
 from typing import Any, Callable
-from .accessibility import AriaExpanded, AriaHasPopup, AriaPressed
-from .events import (
-    ButtonType,
-    FocusEventCallable,
-    KeyboardEventCallable,
-    PressEventCallable,
-    StaticColor,
-    Orientation,
-)
-from .layout import (
+from .types import (
+    # Validation
+    ContextualHelperVariant,
+    # Layout
     AlignSelf,
     CSSProperties,
     DimensionValue,
     JustifySelf,
     LayoutFlex,
-    Number,
+    Placement,
     Position,
-    LabelPosition,
 )
-from .basic import spectrum_element
-from ...elements import Element
+from .basic import component_element
+from ..elements import Element
 
 
-def content(
+def contextual_help(
     *children: Any,
+    variant: ContextualHelperVariant | None = "help",
+    placement: Placement | None = "bottom start",
+    is_open: bool | None = None,
+    default_open: bool | None = None,
+    container_padding: Number | None = None,
+    offset: Number | None = None,
+    cross_offset: Number | None = None,
+    should_flip: bool | None = None,
+    on_open_change: Callable[[bool], None] | None = None,
     flex: LayoutFlex | None = None,
     flex_grow: Number | None = None,
     flex_shrink: Number | None = None,
@@ -34,11 +37,11 @@ def content(
     order: Number | None = None,
     grid_area: str | None = None,
     grid_row: str | None = None,
-    grid_row_start: str | None = None,
-    grid_row_end: str | None = None,
     grid_column: str | None = None,
     grid_column_start: str | None = None,
     grid_column_end: str | None = None,
+    grid_row_start: str | None = None,
+    grid_row_end: str | None = None,
     margin: DimensionValue | None = None,
     margin_top: DimensionValue | None = None,
     margin_bottom: DimensionValue | None = None,
@@ -55,21 +58,33 @@ def content(
     position: Position | None = None,
     top: DimensionValue | None = None,
     bottom: DimensionValue | None = None,
-    start: DimensionValue | None = None,
-    end: DimensionValue | None = None,
     left: DimensionValue | None = None,
     right: DimensionValue | None = None,
+    start: DimensionValue | None = None,
+    end: DimensionValue | None = None,
     z_index: Number | None = None,
     is_hidden: bool | None = None,
     id: str | None = None,
+    aria_label: str | None = None,
+    aria_labelledby: str | None = None,
+    aria_describedby: str | None = None,
+    aria_details: str | None = None,
     UNSAFE_class_name: str | None = None,
     UNSAFE_style: CSSProperties | None = None,
 ) -> Element:
     """
-    Content represents the primary content within a Spectrum container.
-
+    A contextual help is a quiet action button that triggers an informational popover.
     Args:
-        children: The content to render within the container.
+        *children: The children of the contextual help popover.
+        variant: Indicates whether contents are informative or provides helpful guidance.
+        placement: The placement of the popover relative to the action button.
+        is_open: Whether the popover is open by default (controlled).
+        default_open: Whether the popover is open by default (uncontrolled).
+        container_padding: The placement padding that should be applied between the element and its surrounding container.
+        offset: The additional offset applied along the main axis between the element and its anchor element.
+        cross_offset: The additional offset applied along the cross axis between the element and its anchor element.
+        should_flip: Whether the element should flip its orientation when there is insufficient room for it to render completely.
+        on_open_change: Handler that is called when the overlay's open state changes.
         flex: When used in a flex layout, specifies how the element will grow or shrink to fit the space available.
         flex_grow: When used in a flex layout, specifies how the element will grow to fit the space available.
         flex_shrink: When used in a flex layout, specifies how the element will shrink to fit the space available.
@@ -92,27 +107,40 @@ def content(
         margin_x: The margin for the left and right sides of the element.
         margin_y: The margin for the top and bottom sides of the element.
         width: The width of the element.
-        min_width: The minimum width of the element.
-        max_width: The maximum width of the element.
         height: The height of the element.
+        min_width: The minimum width of the element.
         min_height: The minimum height of the element.
+        max_width: The maximum width of the element.
         max_height: The maximum height of the element.
-        position: The position of the element.
-        top: The distance from the top of the containing element.
-        bottom: The distance from the bottom of the containing element.
-        left: The distance from the left of the containing element.
-        right: The distance from the right of the containing element.
-        start: The distance from the start of the containing element, depending on layout direction.
-        end: The distance from the end of the containing element, depending on layout direction.
-        z_index: The stack order of the element.
-        is_hidden: Whether the element is hidden.
+        position: Specifies how the element is position.
+        top: The top position of the element.
+        bottom: The bottom position of the element.
+        left: The left position of the element.
+        right: The right position of the element.
+        start: The logical start position of the element, depending on layout direction.
+        end: The logical end position of the element, depending on layout direction.
+        z_index: The stacking order for the element
+        is_hidden: Hides the element.
         id: The unique identifier of the element.
-        UNSAFE_class_name: A CSS class to apply to the element.
-        UNSAFE_style: A CSS style to apply to the element.
+        aria-label: Defines a string value that labels the current element.
+        aria-labelledby: Identifies the element (or elements) that labels the current element.
+        aria-describedby: Identifies the element (or elements) that describes the object.
+        aria-details: Identifies the element (or elements) that provide a detailed, extended description for the object.
+        UNSAFE_class_name: Set the CSS className for the element. Only use as a last resort. Use style props instead.
+        UNSAFE_style: Set the inline style for the element. Only use as a last resort. Use style props instead.
     """
-    return spectrum_element(
-        "Content",
-        children=children,
+    return component_element(
+        "ContextualHelp",
+        *children,
+        variant=variant,
+        placement=placement,
+        is_open=is_open,
+        default_open=default_open,
+        container_padding=container_padding,
+        offset=offset,
+        cross_offset=cross_offset,
+        should_flip=should_flip,
+        on_open_change=on_open_change,
         flex=flex,
         flex_grow=flex_grow,
         flex_shrink=flex_shrink,
@@ -122,11 +150,11 @@ def content(
         order=order,
         grid_area=grid_area,
         grid_row=grid_row,
-        grid_row_start=grid_row_start,
-        grid_row_end=grid_row_end,
         grid_column=grid_column,
         grid_column_start=grid_column_start,
         grid_column_end=grid_column_end,
+        grid_row_start=grid_row_start,
+        grid_row_end=grid_row_end,
         margin=margin,
         margin_top=margin_top,
         margin_bottom=margin_bottom,
@@ -143,13 +171,17 @@ def content(
         position=position,
         top=top,
         bottom=bottom,
-        start=start,
-        end=end,
         left=left,
         right=right,
+        start=start,
+        end=end,
         z_index=z_index,
         is_hidden=is_hidden,
         id=id,
+        aria_label=aria_label,
+        aria_labelledby=aria_labelledby,
+        aria_describedby=aria_describedby,
+        aria_details=aria_details,
         UNSAFE_class_name=UNSAFE_class_name,
         UNSAFE_style=UNSAFE_style,
     )
