+++ plugins/ui/src/deephaven/ui/components/date_picker.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any, Sequence, Callable
 
-from .spectrum import (
+from .types import (
     FocusEventCallable,
     KeyboardEventCallable,
     LayoutFlex,
@@ -23,13 +23,14 @@ from .spectrum import (
 )
 
 from ..hooks import use_memo
-from ..elements import Element, BaseElement
+from ..elements import Element
 from .._internal.utils import (
     create_props,
     convert_date_props,
     convert_list_prop,
 )
 from ..types import Date, Granularity
+from .basic import component_element
 
 DatePickerElement = Element
 
@@ -271,4 +272,4 @@ def date_picker(
         [unavailable_values],
     )
 
-    return BaseElement("deephaven.ui.components.DatePicker", **props)
+    return component_element("DatePicker", **props)
