+++ plugins/ui/src/deephaven/ui/components/list_view.py
@@ -5,8 +5,9 @@ from typing import Callable, Any
 from deephaven.table import Table
 
 from .item_table_source import ItemTableSource
-from ..elements import BaseElement, Element
+from ..elements import Element
 from .._internal.utils import create_props, unpack_item_table_source
+from .basic import component_element
 from .item import Item
 from ..types import ListViewDensity, Selection, SelectionMode
 
@@ -70,4 +71,4 @@ def list_view(
 
     children, props = unpack_item_table_source(children, props, SUPPORTED_SOURCE_ARGS)
 
-    return BaseElement("deephaven.ui.components.ListView", *children, **props)
+    return component_element("ListView", *children, **props)
