+++ plugins/ui/src/deephaven/ui/components/section.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 from typing import Any
 
 from .._internal.utils import create_props
-from ..elements import Element, BaseElement
+from ..elements import Element
+from .basic import component_element
 from .item import Item
 
 SectionElement = Element
@@ -24,4 +25,4 @@ def section(*children: Item, title: str | None = None, **props: Any) -> SectionE
 
     children, props = create_props(locals())
 
-    return BaseElement("deephaven.ui.components.Section", *children, **props)
+    return component_element("Section", *children, **props)
