+++ plugins/ui/src/deephaven/ui/components/slider.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 from typing import Any, Callable
-
-from .events import (
-    FocusEventCallable,
-    KeyboardEventCallable,
-)
-from .layout import (
+from .types import (
+    # Events
+    Orientation,
+    # Layout
     AlignSelf,
     CSSProperties,
     DimensionValue,
@@ -14,49 +12,31 @@ from .layout import (
     Number,
     Position,
     LabelPosition,
-    Align,
 )
-from .validate import (
-    TextFieldValidationState,
-    NecessityIndicator,
-)
-from .basic import spectrum_element
-from ...elements import Element
+from .basic import component_element
+from ..elements import Element
 
 
-def number_field(
-    is_quiet: bool | None = None,
-    hide_stepper: bool | None = None,
-    decrement_aria_label: str | None = None,
-    increment_aria_label: str | None = None,
-    is_wheel_disabled: bool | None = None,
+def slider(
+    is_filled: bool | None = None,
+    fill_offset: Number | None = None,
+    track_gradient: list[str] | None = None,
     # format_options, # omitted because need to connect it to Deephaven formatting options as well
-    is_disabled: bool | None = None,
-    is_read_only: bool | None = None,
-    is_required: bool | None = None,
-    # validation_behaviour, # omitted because validate is not implemented
-    # validate, # omitted because it needs to return a ValidationError synchronously
-    auto_focus: bool | None = None,
-    value: float | None = None,
-    default_value: float | None = None,
-    min_value: float | None = None,
-    max_value: float | None = None,
-    step: float | None = None,
-    label: Any | None = None,
-    description: Any | None = None,
-    error_message: Any | None = None,
-    validation_state: TextFieldValidationState | None = None,
-    name: str | None = None,
     label_position: LabelPosition = "top",
-    label_align: Align = "start",
-    necessity_indicator: NecessityIndicator = "icon",
+    show_value_label: bool | None = None,
+    # get_value_label, # omitted because it needs to return a string synchronously
     contextual_help: Any | None = None,
-    on_focus: FocusEventCallable | None = None,
-    on_blur: FocusEventCallable | None = None,
-    on_focus_change: Callable[[bool], None] | None = None,
-    on_key_down: KeyboardEventCallable | None = None,
-    on_key_up: KeyboardEventCallable | None = None,
-    on_change: Callable[[float], None] | None = None,
+    orientation: Orientation = "horizontal",
+    is_disabled: bool | None = None,
+    min_value: Number = 0,
+    max_value: Number = 100,
+    step: Number = 1,
+    value: Number | None = None,
+    default_value: Number | None = None,
+    label: Any | None = None,
+    name: str | None = None,
+    on_change_end: Callable[[Number], None] | None = None,
+    on_change: Callable[[Number], None] | None = None,
     flex: LayoutFlex | None = None,
     flex_grow: Number | None = None,
     flex_shrink: Number | None = None,
@@ -100,40 +80,27 @@ def number_field(
     aria_details: str | None = None,
     UNSAFE_class_name: str | None = None,
     UNSAFE_style: CSSProperties | None = None,
-    # missing properties that are clipboard or composition events
 ) -> Element:
     """
-    NumberFields allow users to enter a number, and increment or decrement the value using stepper buttons.
+    Sliders allow users to quickly select a value within a range. They should be used when the upper and lower bounds to the range are invariable.
 
     Args:
-        is_quiet: Whether the input should be displayed with a quiet style
-        hide_stepper: Whether to hide the increment and decrement stepper buttons
-        decrement_aria_label: The aria label for the decrement stepper button. If not provided, the default is "Decrement"
-        increment_aria_label: The aria label for the increment stepper button. If not provided, the default is "Increment"
-        is_wheel_disabled: Whether the input should change with scroll
-        is_disabled: Whether the input should be disabled
-        is_read_only: Whether the input scan be selected but not changed by the user
-        is_required: Whether the input is required before form submission
-        auto_focus: Whether the input should be focused on page load
-        value: The current value of the input
-        default_value: The default value of the input
-        min_value: The minimum value of the input
-        max_value: The maximum value of the input
-        step: The step value for the input
-        label: The label for the input
-        description: A description for the field. Provides a hint such as specific requirements for what to choose.
-        error_message: An error message to display when the field is invalid
-        validation_state: Whether the input should display its "valid" or "invalid" state
-        name: The name of the input, used when submitting an HTML form
-        label_position: The position of the label relative to the input
-        label_align: The alignment of the label relative to the input
-        necessity_indicator: Whether the required state should be shown as an icon or text
-        contextual_help: A ContentualHelp element to place next to the label
-        on_focus: Function called when the button receives focus.
-        on_blur: Function called when the button loses focus.
-        on_focus_change: Function called when the focus state changes.
-        on_key_down: Function called when a key is pressed.
-        on_key_up: Function called when a key is released.
+        is_filled: Whether the slider should be filled between the start of the slider and the current value.
+        fill_offset: The offset of the filled area from the start of the slider.
+        track_gradient: The background of the track, specified as the stops for a CSS background: linear-gradient(to right/left, ...trackGradient).
+        label_position: The position of the label relative to the slider.
+        show_value_label: Whether the value label should be displayed. True by default if the label is provided.
+        contextual_help: A ContextualHelp element to place next to the label.
+        orientation: The orientation of the slider.
+        is_disabled: Whether the slider is disabled.
+        min_value: The minimum value of the slider.
+        max_value: The maximum value of the slider.
+        step: The step value for the slider.
+        value: The current value of the slider.
+        default_value: The default value of the slider.
+        label: The content to display as the label.
+        name: The name of the input element, used when submitting an HTML form.
+        on_change_end: Function called when the slider stops moving
         on_change: Function called when the input value changes
         flex: When used in a flex layout, specifies how the element will grow or shrink to fit the space available.
         flex_grow: When used in a flex layout, specifies how the element will grow to fit the space available.
@@ -179,37 +146,26 @@ def number_field(
         UNSAFE_class_name: A CSS class to apply to the element.
         UNSAFE_style: A CSS style to apply to the element.
     """
-
-    return spectrum_element(
-        "NumberField",
-        is_quiet=is_quiet,
-        hide_stepper=hide_stepper,
-        decrement_aria_label=decrement_aria_label,
-        increment_aria_label=increment_aria_label,
-        is_wheel_disabled=is_wheel_disabled,
+    return component_element(
+        "Slider",
+        is_filled=is_filled,
+        fill_offset=fill_offset,
+        track_gradient=track_gradient,
+        # format_options=format_options,
+        label_position=label_position,
+        show_value_label=show_value_label,
+        # get_value_label=get_value_label,
+        contextual_help=contextual_help,
+        orientation=orientation,
         is_disabled=is_disabled,
-        is_read_only=is_read_only,
-        is_required=is_required,
-        auto_focus=auto_focus,
-        value=value,
-        default_value=default_value,
         min_value=min_value,
         max_value=max_value,
         step=step,
+        value=value,
+        default_value=default_value,
         label=label,
-        description=description,
-        error_message=error_message,
-        validation_state=validation_state,
         name=name,
-        label_position=label_position,
-        label_align=label_align,
-        necessity_indicator=necessity_indicator,
-        contextual_help=contextual_help,
-        on_focus=on_focus,
-        on_blur=on_blur,
-        on_focus_change=on_focus_change,
-        on_key_down=on_key_down,
-        on_key_up=on_key_up,
+        on_change_end=on_change_end,
         on_change=on_change,
         flex=flex,
         flex_grow=flex_grow,
