--- plugins/ui/src/deephaven/ui/components/spectrum/__init__.py
@@ -1,16 +1,0 @@
-from .action_button import *
-from .basic import *
-from .button_group import *
-from .button import *
-from .checkbox import *
-from .content import *
-from .contextual_help import *
-from .number_field import *
-from .range_slider import *
-from .slider import *
-from .text_field import *
-from .toggle_button import *
-from .flex import *
-from .date_picker import *
-from .combo_box import *
-from .layout import *
