+++ plugins/ui/src/deephaven/ui/components/text_field.py
@@ -1,15 +1,13 @@
 from __future__ import annotations
 from typing import Any, Callable
-from .accessibility import AriaExpanded, AriaHasPopup, AriaPressed
-from .events import (
-    ButtonType,
+from .types import (
+    # Accessibility
+    AriaHasPopup,
+    AriaAutoComplete,
+    # Events
     FocusEventCallable,
     KeyboardEventCallable,
-    PressEventCallable,
-    StaticColor,
-    Orientation,
-)
-from .layout import (
+    # Layout
     AlignSelf,
     CSSProperties,
     DimensionValue,
@@ -18,31 +16,47 @@ from .layout import (
     Number,
     Position,
     LabelPosition,
+    Align,
+    # Validation
+    TextFieldType,
+    TextFieldInputMode,
+    TextFieldValidationState,
+    NecessityIndicator,
 )
-from .basic import spectrum_element
-from ...elements import Element
+from .basic import component_element
+from ..elements import Element
 
 
-def slider(
-    is_filled: bool | None = None,
-    fill_offset: Number | None = None,
-    track_gradient: list[str] | None = None,
-    # format_options, # omitted because need to connect it to Deephaven formatting options as well
-    label_position: LabelPosition = "top",
-    show_value_label: bool | None = None,
-    # get_value_label, # omitted because it needs to return a string synchronously
-    contextual_help: Any | None = None,
-    orientation: Orientation = "horizontal",
+def text_field(
+    icon: Element | None = None,
+    is_quiet: bool | None = None,
     is_disabled: bool | None = None,
-    min_value: Number = 0,
-    max_value: Number = 100,
-    step: Number = 1,
-    value: Number | None = None,
-    default_value: Number | None = None,
+    is_read_only: bool | None = None,
+    is_required: bool | None = None,
+    description: Any | None = None,
+    error_message: Any | None = None,
+    auto_focus: bool | None = None,
+    value: str | None = None,
+    default_value: str | None = None,
     label: Any | None = None,
+    auto_complete: str | None = None,
+    max_length: int | None = None,
+    min_length: int | None = None,
+    pattern: str | None = None,
+    type: TextFieldType | None = None,
+    input_mode: TextFieldInputMode | None = None,
     name: str | None = None,
-    on_change_end: Callable[[Number], None] | None = None,
-    on_change: Callable[[Number], None] | None = None,
+    validation_state: TextFieldValidationState | None = None,
+    label_position: LabelPosition = "top",
+    label_align: Align = "start",
+    necessity_indicator: NecessityIndicator = "icon",
+    contextual_help: Any | None = None,
+    on_focus: FocusEventCallable | None = None,
+    on_blur: FocusEventCallable | None = None,
+    on_focus_change: Callable[[bool], None] | None = None,
+    on_key_down: KeyboardEventCallable | None = None,
+    on_key_up: KeyboardEventCallable | None = None,
+    on_change: Callable[[str], None] | None = None,
     flex: LayoutFlex | None = None,
     flex_grow: Number | None = None,
     flex_shrink: Number | None = None,
@@ -80,33 +94,51 @@ def slider(
     z_index: Number | None = None,
     is_hidden: bool | None = None,
     id: str | None = None,
+    exclude_from_tab_order: bool | None = None,
+    aria_active_descendant: str | None = None,
+    aria_auto_complete: AriaAutoComplete | None = None,
+    aria_haspopup: AriaHasPopup | None = None,
     aria_label: str | None = None,
     aria_labelledby: str | None = None,
     aria_describedby: str | None = None,
     aria_details: str | None = None,
+    aria_errormessage: str | None = None,
     UNSAFE_class_name: str | None = None,
     UNSAFE_style: CSSProperties | None = None,
+    # missing properties that are clipboard or composition events
 ) -> Element:
     """
-    Sliders allow users to quickly select a value within a range. They should be used when the upper and lower bounds to the range are invariable.
+    TextFields are text inputs that allow users to input custom text entries with a keyboard. Various decorations can be displayed around the field to communicate the entry requirements.
 
     Args:
-        is_filled: Whether the slider should be filled between the start of the slider and the current value.
-        fill_offset: The offset of the filled area from the start of the slider.
-        track_gradient: The background of the track, specified as the stops for a CSS background: linear-gradient(to right/left, ...trackGradient).
-        label_position: The position of the label relative to the slider.
-        show_value_label: Whether the value label should be displayed. True by default if the label is provided.
-        contextual_help: A ContextualHelp element to place next to the label.
-        orientation: The orientation of the slider.
-        is_disabled: Whether the slider is disabled.
-        min_value: The minimum value of the slider.
-        max_value: The maximum value of the slider.
-        step: The step value for the slider.
-        value: The current value of the slider.
-        default_value: The default value of the slider.
-        label: The content to display as the label.
-        name: The name of the input element, used when submitting an HTML form.
-        on_change_end: Function called when the slider stops moving
+        icon: An icon to display at the start of the input
+        is_quiet: Whether the input should be displayed with a quiet style
+        is_disabled: Whether the input should be disabled
+        is_read_only: Whether the input scan be selected but not changed by the user
+        is_required: Whether the input is required before form submission
+        description: A description for the field. Provides a hint such as specific requirements for what to choose.
+        error_message: An error message to display when the field is invalid
+        auto_focus: Whether the input should be focused on page load
+        value: The current value of the input
+        default_value: The default value of the input
+        label: The label for the input
+        auto_complete: Describes the type of autocomplete functionality the input should provide
+        max_length: The maximum number of characters the input can accept
+        min_length: The minimum number of characters the input can accept
+        pattern: A regular expression that the input's value must match
+        type: The type of input to display
+        input_mode: Hints at the tpye of data that might be entered by the user while editing the element or its contents
+        name: The name of the input, used when submitting an HTML form
+        validation_state: Whether the input should display its "valid" or "invalid" state
+        label_position: The position of the label relative to the input
+        label_align: The alignment of the label relative to the input
+        necessity_indicator: Whether the required state should be shown as an icon or text
+        contextual_help: A ContentualHelp element to place next to the label
+        on_focus: Function called when the button receives focus.
+        on_blur: Function called when the button loses focus.
+        on_focus_change: Function called when the focus state changes.
+        on_key_down: Function called when a key is pressed.
+        on_key_up: Function called when a key is released.
         on_change: Function called when the input value changes
         flex: When used in a flex layout, specifies how the element will grow or shrink to fit the space available.
         flex_grow: When used in a flex layout, specifies how the element will grow to fit the space available.
@@ -145,33 +177,48 @@ def slider(
         z_index: The stack order of the element.
         is_hidden: Whether the element is hidden.
         id: The unique identifier of the element.
+        exclude_from_tab_order: Whether the element should be excluded from the tab order.
+        aria_active_descendant: Identifies the currently active element when DOM focus is on a composite widget, textbox, group, or application.
+        aria_auto_complete: Indicates whether inputting text could trigger display of one or more predictions of the user's intended value for an input and specifies how predictions would be presented if they are made.
         aria_label: The label for the element.
         aria_labelled_by: The id of the element that labels the current element.
         aria_described_by: The id of the element that describes the current element.
         aria_details: The id of the element that provides additional information about the current element.
+        aria_errormessage: The id of the element that provides an error message for the current element.
         UNSAFE_class_name: A CSS class to apply to the element.
         UNSAFE_style: A CSS style to apply to the element.
     """
-    return spectrum_element(
-        "Slider",
-        is_filled=is_filled,
-        fill_offset=fill_offset,
-        track_gradient=track_gradient,
-        # format_options=format_options,
-        label_position=label_position,
-        show_value_label=show_value_label,
-        # get_value_label=get_value_label,
-        contextual_help=contextual_help,
-        orientation=orientation,
+
+    return component_element(
+        "TextField",
+        icon=icon,
+        is_quiet=is_quiet,
         is_disabled=is_disabled,
-        min_value=min_value,
-        max_value=max_value,
-        step=step,
+        is_read_only=is_read_only,
+        is_required=is_required,
+        description=description,
+        error_message=error_message,
+        auto_focus=auto_focus,
         value=value,
         default_value=default_value,
         label=label,
+        auto_complete=auto_complete,
+        max_length=max_length,
+        min_length=min_length,
+        pattern=pattern,
+        type=type,
+        input_mode=input_mode,
         name=name,
-        on_change_end=on_change_end,
+        validation_state=validation_state,
+        label_position=label_position,
+        label_align=label_align,
+        necessity_indicator=necessity_indicator,
+        contextual_help=contextual_help,
+        on_focus=on_focus,
+        on_blur=on_blur,
+        on_focus_change=on_focus_change,
+        on_key_down=on_key_down,
+        on_key_up=on_key_up,
         on_change=on_change,
         flex=flex,
         flex_grow=flex_grow,
@@ -210,10 +257,15 @@ def slider(
         z_index=z_index,
         is_hidden=is_hidden,
         id=id,
+        exclude_from_tab_order=exclude_from_tab_order,
+        aria_active_descendant=aria_active_descendant,
+        aria_auto_complete=aria_auto_complete,
+        aria_haspopup=aria_haspopup,
         aria_label=aria_label,
         aria_labelledby=aria_labelledby,
         aria_describedby=aria_describedby,
         aria_details=aria_details,
+        aria_errormessage=aria_errormessage,
         UNSAFE_class_name=UNSAFE_class_name,
         UNSAFE_style=UNSAFE_style,
     )
