+++ tests/app.d/tests.app
@@ -6,3 +6,4 @@ name=Plugins Test Application
 file_0=express.py
 file_1=matplotlib.py
 file_2=ui.py
+file_3=ui_render_all.py
